refactor: move function `configureExecModule` to `exec.ts`
To align the code and file structure with the `container` module.
