+++ core/test/unit/src/plugins/exec/exec.ts
@@ -12,7 +12,7 @@ import { join } from "path"
 import psTree from "ps-tree"
 
 import { Garden } from "../../../../../src/garden"
-import { ExecProvider, gardenPlugin, getLogFilePath } from "../../../../../src/plugins/exec/exec"
+import { configureExecModule, ExecProvider, gardenPlugin, getLogFilePath } from "../../../../../src/plugins/exec/exec"
 import { LogEntry } from "../../../../../src/logger/log-entry"
 import { keyBy } from "lodash"
 import {
@@ -33,7 +33,7 @@ import { TestTask } from "../../../../../src/tasks/test"
 import { readFile, remove } from "fs-extra"
 import { dedent } from "../../../../../src/util/string"
 import { sleep } from "../../../../../src/util/util"
-import { configureExecModule, ExecModuleConfig } from "../../../../../src/plugins/exec/moduleConfig"
+import { ExecModuleConfig } from "../../../../../src/plugins/exec/moduleConfig"
 import { actionFromConfig } from "../../../../../src/graph/actions"
 import { TestAction, TestActionConfig } from "../../../../../src/actions/test"
 import { PluginContext } from "../../../../../src/plugin-context"
